review_features.zip
A csv containing feature sets for all the reviews in our data set. The features are based on the count of the 1000 most common words across the entire data set (which was stripped of stop words and single characters). It does not contain tf-idf scores as originally planned, as the run-time to generate this data was excessive.